feat: 공통 API 통신 기반과 그룹 계약 모델 추가 (#424)
* feat: 공통 API 통신 기반 추가

* feat: 그룹 API 계약 모델 추가

* feat: 그룹 API 모델의 화면 매핑 추가

* fix: 그룹 API 응답 필드 검증

+++ AN/Pheeeew/shared/src/commonMain/kotlin/com/pheeeew/feature/screens/group/model/GroupSummaryUiModel.kt
@@ -6,7 +6,7 @@ import com.pheeeew.feature.component.stamp.StampAppearanceUiModel
 data class GroupSummaryUiModel(
     val id: GroupId,
     val name: String,
-    val memberCount: Int,
+    val memberCount: Long,
     val weeklyStampCount: Long,
     val stamp: StampAppearanceUiModel,
 ) {
